add visualize.py
